BAM! Concurrent repo processing. For real.
No sleeping, no waiting, no dead messages. I can parse all the repos I
want and have the dispatcher decide when it's appropriate to shutdown
the Actor system.

There is one caveat- I know when to trigger the shutdown of the system,
but I'm not able to shutdown the FileCreator Actors as they report that
they've written a file. I have to wait until everything's done and then
kill the system. Ideally I would be able to shut those guys down as they
finish.