Add sealed SubCommand trait to Client trait
This is a strange idea for me, and I'm not sure what benefits it will
actually provide.

I plan to have all subcommands that a Client supports be contained by
this trait. It would be simpler to have them stored as a collection of
strings, but I'm hoping this leads to some more interesting places.